+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Fixes #PAC-348: Prevent processing global attributes in all stores
   Make Actions and ActionInterfaces deprecated, replace DI configuration with GenericAction + GenericIdentifierAction
 * Prepare generic workflow and defined deprecated interface PrimaryKeyUtilInterface
+* PAC-362: Call to a member function getSystemLogger() on null
   
 ## Features